+++ src/main/java/com/b1n_ry/yigd/networking/ClientPacketHandler.java
@@ -6,6 +6,7 @@ import com.b1n_ry.yigd.client.gui.PlayerSelectionGui;
 import com.b1n_ry.yigd.client.gui.screens.GraveOverviewScreen;
 import com.b1n_ry.yigd.client.gui.screens.GraveSelectionScreen;
 import com.b1n_ry.yigd.client.gui.screens.PlayerSelectionScreen;
+import com.b1n_ry.yigd.client.render.GraveBlockEntityRenderer;
 import com.b1n_ry.yigd.components.GraveComponent;
 import com.b1n_ry.yigd.config.YigdConfig;
 import com.b1n_ry.yigd.networking.packets.*;
@@ -18,7 +19,7 @@ import java.util.UUID;
 
 public class ClientPacketHandler {
     public static void registerReceivers() {
-        ClientPlayNetworking.registerGlobalReceiver(GraveOverviewS2CPacket.ID, (payload, context) -> {
+        ClientPlayNetworking.registerGlobalReceiver(GraveOverviewS2CPacket.TYPE, (payload, context) -> {
             GraveComponent component = payload.component();
             boolean canRestore = payload.canRestore();
             boolean canRob = payload.canRob();
@@ -33,18 +34,24 @@ public class ClientPacketHandler {
                     client.screen, canRestore, canRob, canDelete, canUnlock,
                     obtainableKeys, obtainableCompass))));
         });
-        ClientPlayNetworking.registerGlobalReceiver(GraveSelectionS2CPacket.ID, (payload, context) -> {
+        ClientPlayNetworking.registerGlobalReceiver(GraveSelectionS2CPacket.TYPE, (payload, context) -> {
             List<LightGraveData> data = payload.data();
             ResolvableProfile profile = payload.owner();
 
             Minecraft client = context.client();
             client.execute(() -> client.setScreen(new GraveSelectionScreen(new GraveSelectionGui(data, profile, client.screen))));
         });
-        ClientPlayNetworking.registerGlobalReceiver(PlayerSelectionS2CPacket.ID, (payload, context) -> {
+        ClientPlayNetworking.registerGlobalReceiver(PlayerSelectionS2CPacket.TYPE, (payload, context) -> {
             List<LightPlayerData> data = payload.data();
 
             Minecraft client = context.client();
             client.execute(() -> client.setScreen(new PlayerSelectionScreen(new PlayerSelectionGui(data, client.screen))));
+        });
+        ClientPlayNetworking.registerGlobalReceiver(SyncConfigS2CPacket.TYPE, (payload, context) -> {
+            YigdConfig.getConfig().graveConfig.retrieveMethods.onBreak = payload.gravesBreakable();
+            GraveBlockEntityRenderer.syncedGlowing = payload.glowingGraves();
+            GraveBlockEntityRenderer.syncedGlowingMaxDistance = payload.maxGraveGlowingDistance();
+            GraveBlockEntityRenderer.syncedDeathSightDistance = payload.deathSightDistance();
         });
     }