+++ src/main/java/com/b1n_ry/yigd/networking/PacketInitializer.java
@@ -5,18 +5,19 @@ import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
 
 public class PacketInitializer {
     public static void init() {
-        PayloadTypeRegistry.playC2S().register(DeleteGraveC2SPacket.ID, DeleteGraveC2SPacket.CODEC);
-        PayloadTypeRegistry.playC2S().register(GraveOverviewRequestC2SPacket.ID, GraveOverviewRequestC2SPacket.CODEC);
-        PayloadTypeRegistry.playC2S().register(GraveSelectionRequestC2SPacket.ID, GraveSelectionRequestC2SPacket.CODEC);
-        PayloadTypeRegistry.playC2S().register(LockGraveC2SPacket.ID, LockGraveC2SPacket.CODEC);
-        PayloadTypeRegistry.playC2S().register(RequestCompassC2SPacket.ID, RequestCompassC2SPacket.CODEC);
-        PayloadTypeRegistry.playC2S().register(RequestKeyC2SPacket.ID, RequestKeyC2SPacket.CODEC);
-        PayloadTypeRegistry.playC2S().register(RestoreGraveC2SPacket.ID, RestoreGraveC2SPacket.CODEC);
-        PayloadTypeRegistry.playC2S().register(RobGraveC2SPacket.ID, RobGraveC2SPacket.CODEC);
-        PayloadTypeRegistry.playC2S().register(UpdateConfigC2SPacket.ID, UpdateConfigC2SPacket.CODEC);
+        PayloadTypeRegistry.playC2S().register(DeleteGraveC2SPacket.TYPE, DeleteGraveC2SPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playC2S().register(GraveOverviewRequestC2SPacket.TYPE, GraveOverviewRequestC2SPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playC2S().register(GraveSelectionRequestC2SPacket.TYPE, GraveSelectionRequestC2SPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playC2S().register(LockGraveC2SPacket.TYPE, LockGraveC2SPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playC2S().register(RequestCompassC2SPacket.TYPE, RequestCompassC2SPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playC2S().register(RequestKeyC2SPacket.TYPE, RequestKeyC2SPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playC2S().register(RestoreGraveC2SPacket.TYPE, RestoreGraveC2SPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playC2S().register(RobGraveC2SPacket.TYPE, RobGraveC2SPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playC2S().register(UpdateConfigC2SPacket.TYPE, UpdateConfigC2SPacket.STREAM_CODEC);
 
-        PayloadTypeRegistry.playS2C().register(GraveOverviewS2CPacket.ID, GraveOverviewS2CPacket.CODEC);
-        PayloadTypeRegistry.playS2C().register(GraveSelectionS2CPacket.ID, GraveSelectionS2CPacket.CODEC);
-        PayloadTypeRegistry.playS2C().register(PlayerSelectionS2CPacket.ID, PlayerSelectionS2CPacket.CODEC);
+        PayloadTypeRegistry.playS2C().register(GraveOverviewS2CPacket.TYPE, GraveOverviewS2CPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playS2C().register(GraveSelectionS2CPacket.TYPE, GraveSelectionS2CPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playS2C().register(PlayerSelectionS2CPacket.TYPE, PlayerSelectionS2CPacket.STREAM_CODEC);
+        PayloadTypeRegistry.playS2C().register(SyncConfigS2CPacket.TYPE, SyncConfigS2CPacket.STREAM_CODEC);
     }
 }