+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # You're in Grave Danger 2.4.14
 
+### Changes
+* Config values for glowing graves, glowing grave distance, breakable graves, and
+death sight enchantment range are now properly synced to the server
+
 ---
 
 # You're in Grave Danger 2.4.13